refactor: useMovieQuery 분리
